Move generated manifests into manifest
