+++ dotpkg/legacy_manifest.py
