+++ dotpkg/manifest/installs.py
@@ -3,15 +3,15 @@ from dataclasses import dataclass
 from typing import Any
 
 @dataclass
-class Installs:
+class InstallsManifest:
     '''A manifest keeping track of the installed locations of dotpkgs'''
     
     version: int
     '''The version of the install manifest.'''
     
     @staticmethod
-    def from_dict(d: dict[str, Any]) -> Installs:
-        return Installs(
+    def from_dict(d: dict[str, Any]) -> InstallsManifest:
+        return InstallsManifest(
             version=d['version'],
         )
